Fit a sparse multivariate regression model using variational inference.

Variational approximation procedure fitting a sparse multivariate regression model for combined selection of predictors and associated responses in high-dimensional set-ups. Dependence across responses linked to the same predictors is captured through the model hierarchical structure.

locus_horseshoe(Y, X, p0_av, Z = NULL, list_hyper = NULL,
  list_init = NULL, list_cv = NULL, list_blocks = NULL,
  user_seed = NULL, tol = 1e-04, maxit = 1000, batch = TRUE,
  save_hyper = FALSE, save_init = FALSE, verbose = TRUE)

Arguments

Y
Continuous response data matrix of dimension n x d, where n is the number of observations and d is the number of response variables.

X
Input matrix of dimension n x p, where p is the number of candidate predictors. X cannot contain NAs. No intercept must be supplied.

Z
Covariate matrix of dimension n x q, where q is the number of covariates. NULL if no covariate. Factor covariates must be supplied after transformation to dummy coding. No intercept must be supplied.

list_hyper
An object of class "hyper" containing the model hyperparameters. Must be filled using the set_hyper function or must be NULL for default hyperparameters.

list_init
An object of class "init" containing the initial variational parameters. Must be filled using the set_init function or be NULL for a default initialization.

list_cv
An object of class "cv" containing settings for choosing the prior average number of predictors expected to be included in the model, p0_av, by cross-validation. Must be filled using the set_cv function or must be NULL for no cross-validation. If non-NULL, p0_av, list_init and list_hyper must all be NULL.

list_blocks
An object of class "blocks" containing setting for parallel inference on a partitioned predictor space. Must be filled using the set_blocks function or must be NULL for no partitioning.

user_seed
Seed set for reproducible default choices of hyperparameters (if list_hyper is NULL) and inital variational parameters (if list_init is NULL). Also used at the cross-validation stage (if list_cv is non-NULL). Default is NULL, no seed set.

tol
Tolerance for the stopping criterion.

maxit
Maximum number of iterations allowed.

batch
If TRUE a fast batch updating scheme is used (recommended).

save_hyper
If TRUE, the hyperparameters used for the model are saved as output.

save_init
If TRUE, the initial variational parameters used for the inference are saved as output.

verbose
If TRUE, messages are displayed during execution.

Value

An object of class "vb" containing the following variational estimates and settings:

lb_opt

Optimized variational lower bound for the marginal log-likelihood.

gam_vb

Posterior inclusion probability matrix of dimension p x d. Entry (s, t) corresponds to the posterior probability of association between predictor s and response t.

mu_alpha_vb

Matrix of dimension q x d whose entries are the posterior mean regression coefficients for the covariates provided in Z. NULL if Z is NULL.

om_vb

Vector of size p containing the posterior mean of omega. Entry s controls the proportion of responses associated with predictor s.

x_prpnst

Vector of size p containing the sums over the colums of gam_vb, used to evaluate the propensity for the candidate predictors to be involved in associations.

y_prpnst

Vector of size d containing the sums over the rows of gam_vb, used to evaluate the propensity for the responses to be involved in associations.

rmvd_cst_x, rmvd_cst_z

Vectors containing the indices of constant variables in X (resp. Z) removed prior to the analysis.

rmvd_coll_x, rmvd_coll_z

Vectors containing the indices of variables in X (resp. Z) removed prior to the analysis because collinear to other variables. The entry name indicates the corresponding variable kept in the analysis (i.e., that causing the collinearity for the entry in question).

list_hyper, list_init

If save_hyper, resp. save_init, TRUE, hyperparameters, resp. initial variational parameters, used for inference are saved as output.

See also

set_hyper, set_init, set_cv, set_blocks

Examples

user_seed <- 123; set.seed(user_seed)
n <- 200; p <- 300; p0 <- 50; d <- 40; d0 <- 30
list_X <- generate_snps(n = n, p = p)
list_Y <- generate_phenos(n = n, d = d, var_err = 0.25)

dat <- generate_dependence(list_snps = list_X, list_phenos = list_Y,
                           ind_d0 = sample(1:d, d0), ind_p0 = sample(1:p, p0),
                           vec_prob_sh = 0.1, max_tot_pve = 0.9)

vb <- locus(Y = dat$phenos, X = dat$snps, p0_av = p0, user_seed = user_seed)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -7791.264
#> 
#> Iteration 5... 
#> Lower bound = -6738.198
#> 
#> Iteration 10... 
#> Lower bound = -6736.132
#> 
#> Convergence obtained after 14 iterations with variational lower bound = -6736.127. 
#> 
#> ... done. == 
#>