Gather settings for the cross-validation procedure used in locus.

The cross-validation procedure uses the variational lower bound as objective function and is used to select the prior average number of predictors p0_av expected to be included in the model. p0_av is used to set the model hyperparameters and ensure sparse predictor selections.

set_cv(n, p, n_folds, size_p0_av_grid, n_cpus, tol_cv = 0.001,
  maxit_cv = 1000, batch_cv = TRUE, verbose = TRUE)

Arguments

n
Number of observations.

p
Number of candidate predictors.

n_folds
Number of number of folds. Large folds are not recommended for large datasets as the procedure may become computationally expensive. Must be greater than 2 and smaller than the number of observations.

size_p0_av_grid
Number of possible values of p0_av to be compared. Large numbers are not recommended for large datasets as the procedure may become computationally expensive.

n_cpus
Number of CPUs to be used for the cross-validation procedure. If large, one should ensure that enough RAM will be available for parallel execution. Set to 1 for serial execution.

tol_cv
Tolerance for the variational algorithm stopping criterion used within the cross-validation procedure.

maxit_cv
Maximum number of iterations allowed for the variational algorithm used within the cross-validation procedure.

batch_cv
If TRUE, a fast batch updating scheme is used within the cross-validation procedure (recommended).

verbose
If TRUE, messages are displayed when calling set_cv.

Value

An object of class "cv" preparing the settings for the cross-validation settings in a form that can be passed to the locus function.

Details

This cross-validation procedure is available only for link = "identity".

See also

locus

Examples

user_seed <- 123; set.seed(user_seed)
n <- 150; p <- 200; p0 <- 50; d <- 25; d0 <- 20
list_X <- generate_snps(n = n, p = p)
list_Y <- generate_phenos(n = n, d = d, var_err = 0.25)

dat <- generate_dependence(list_snps = list_X, list_phenos = list_Y,
                           ind_d0 = sample(1:d, d0),
                           ind_p0 = sample(1:p, p0), vec_prob_sh = 0.1,
                           max_tot_pve = 0.9)

list_cv <- set_cv(n, p, n_folds = 3, size_p0_av_grid = 3, n_cpus = 2)

#> Cross-validation with 2 CPUs.
#> Please make sure that enough RAM is available. 

vb <- locus(Y = dat$phenos, X = dat$snps, p0_av = NULL, link = "identity",
            list_cv = list_cv, user_seed = user_seed)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> =============================== 
#> ===== Cross-validation... ===== 
#> =============================== 
#> Lower bounds on test sets for each fold and each grid element: 
#>         p0_av_50  p0_av_75 p0_av_100
#> fold_1 -1530.083 -1583.555 -1658.641
#> fold_2 -1532.297 -1560.215 -1604.408
#> fold_3 -1547.369 -1574.037 -1610.658
#> ===== ...end of cross-validation with selected p0_av = 50 ===== 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -3943.702
#> 
#> Iteration 5... 
#> Lower bound = -3316.677
#> 
#> Iteration 10... 
#> Lower bound = -3310.55
#> 
#> Iteration 15... 
#> Lower bound = -3310.498
#> 
#> Iteration 20... 
#> Lower bound = -3310.451
#> 
#> Iteration 25... 
#> Lower bound = -3310.444
#> 
#> Convergence obtained after 27 iterations with variational lower bound = -3310.444. 
#> 
#> ... done. == 
#>