Gather model hyperparameters provided by the user.

This function must be used to provide hyperparameter values for the model used in locus.

set_hyper(d, p, lambda, nu, a, b, eta, kappa, link = "identity",
  ind_bin = NULL, q = NULL, phi = NULL, xi = NULL, r = NULL,
  m0 = NULL)

Arguments

d
Number of responses.

p
Number of candidate predictors.

lambda
Vector of length 1 providing the values of hyperparameter \(\lambda\) for the prior distribution of \(\sigma^{-2}\). \(\sigma\) represents the typical size of nonzero effects.

nu
Vector of length 1 providing the values of hyperparameter \(\nu\) for the prior distribution of \(\sigma^{-2}\). \(\sigma\) represents the typical size of nonzero effects.

a
Vector of length 1 or p providing the values of hyperparameter \(a\) for the prior distributions for the proportion of responses associated with each candidate predictor, \(\omega\) (vector of length p). If of length 1, the provided value is repeated p times.

b
Vector of length 1 or p providing the values of hyperparameter \(b\) for the prior distributions for the proportion of responses associated with each candidate predictor, \(\omega\) (vector of length p). If of length 1, the provided value is repeated p times.

eta
Vector of length 1 or d for link = "identity", and of length 1 or d_cont = d - length(ind_bin) (the number of continuous response variables) for link = "mix". Provides the values of hyperparameter \(\eta\) for the prior distributions of the continuous response residual precisions, \(\tau\). If of length 1, the provided value is repeated d, resp. d_cont, times. Must be NULL for link = "logit" and link = "probit".

kappa
Vector of length 1 or d for link = "identity", and of length 1 or d_cont = d - length(ind_bin) (the number of continuous response variables) for link = "mix". Provides the values of hyperparameter \(\kappa\) for the prior distributions of the response residual precisions, \(\tau\). If of length 1, the provided value is repeated d, resp. d_cont, times. Must be NULL for link = "logit" and link = "probit".

link
Response link. Must be "identity" for linear regression, "logit" for logistic regression, "probit" for probit regression, or "mix" for a mix of identity and probit link functions (in this case, the indices of the binary responses must be gathered in argument ind_bin, see below).

ind_bin
If link = "mix", vector of indices corresponding to the binary variables in Y. Must be NULL if link != "mix".

q
Number of covariates. Default is NULL, for Z NULL.

phi
Vector of length 1 or q providing the values of hyperparameter \(\phi\) for the prior distributions for the sizes of the nonzero covariate effects, \(\zeta\). If of length 1, the provided value is repeated q times. Default is NULL, for Z NULL.

xi
Vector of length 1 or q providing the values of hyperparameter \(\xi\) for the prior distributions for the sizes of the nonzero covariate effects, \(\zeta\). If of length 1, the provided value is repeated q times. Default is NULL, for Z NULL.

r
Number of variables representing external information on the candidate predictors. Default is NULL, for V NULL.

m0
Vector of length 1 or p providing the values of hyperparameter \(m0\) for the prior distribution of \(c0\) linked to the proportion of responses associated with each candidate predictor. If of length 1, the provided value is repeated p times. Default is NULL, for V NULL.

Value

An object of class "hyper" preparing user hyperparameter in a form that can be passed to the locus function.

Details

The locus function can also be used with default hyperparameter choices (without using set_hyper) by setting its argument list_hyper to NULL.

See also

set_init, locus

Examples

user_seed <- 123
n <- 200; p <- 250; p0 <- 50; d <- 25; d0 <- 15
list_X <- generate_snps(n = n, p = p, user_seed = user_seed)
list_Y <- generate_phenos(n = n, d = d, var_err = 1, user_seed = user_seed)

# Continuous outcomes
#
dat_g <- generate_dependence(list_snps = list_X, list_phenos = list_Y,
                             ind_d0 = sample(1:d, d0),
                             ind_p0 = sample(1:p, p0), vec_prob_sh = 0.1,
                             family = "gaussian", max_tot_pve = 0.9,
                             user_seed = user_seed)

# a and b chosen so that the prior mean number of responses associated with
# each candidate predictor is 1/4.
list_hyper_g <- set_hyper(d, p, lambda = 1, nu = 1, a = 1, b = 4*d-1,
                          eta = 1, kappa = apply(dat_g$phenos, 2, var),
                          link = "identity")

# we take p0_av = p0 (known here); this choice may result in variable
# selections that are (too) conservative in some cases. In practice, often
# p0_av as a slightly overestimated guess of p0.
vb_g <- locus(Y = dat_g$phenos, X = dat_g$snps, p0_av = p0,
              link = "identity", list_hyper = list_hyper_g,
              user_seed = user_seed)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -8960.294
#> 
#> Iteration 5... 
#> Lower bound = -7851.613
#> 
#> Convergence obtained after 10 iterations with variational lower bound = -7851.393. 
#> 
#> ... done. == 
#> 

# Continuous outcomes with covariates
#
q <- 4
Z <- matrix(rnorm(n * q), nrow = n)

list_hyper_g_z <- set_hyper(d, p, lambda = 1, nu = 1, a = 1, b = 4*d-1,
                            eta = 1, kappa = apply(dat_g$phenos, 2, var),
                            link = "identity", q = q, phi = 1, xi = 1)

vb_g_z <- locus(Y = dat_g$phenos, X = dat_g$snps, p0_av = p0, Z = Z,
                link = "identity", list_hyper = list_hyper_g_z,
                user_seed = user_seed)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -12527.34
#> 
#> Iteration 5... 
#> Lower bound = -7988.895
#> 
#> Iteration 10... 
#> Lower bound = -7984.033
#> 
#> Convergence obtained after 12 iterations with variational lower bound = -7984.033. 
#> 
#> ... done. == 
#> 

# Continuous outcomes with external annotation
#
r <- 4
V <- matrix(rnorm(p * r), nrow = p)
bool_p0 <- rowSums(dat_g$pat) > 0
V[bool_p0, ] <- rnorm(sum(bool_p0) * r, mean = 2) # informative annotations

list_hyper_g_v <- set_hyper(d, p, lambda = 1, nu = 1, a = NULL, b = NULL,
                            eta = 1, kappa = apply(dat_g$phenos, 2, var),
                            link = "identity", r = r, m0 = 0)

vb_g_v <- locus(Y = dat_g$phenos, X = dat_g$snps, p0_av = p0,  V = V,
                link = "identity", list_hyper = list_hyper_g_v,
                user_seed = user_seed)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -19079.82
#> 
#> Iteration 5... 
#> Lower bound = -9371.355
#> 
#> Iteration 10... 
#> Lower bound = -8743.522
#> 
#> Iteration 15... 
#> Lower bound = -8645.668
#> 
#> Iteration 20... 
#> Lower bound = -8619.123
#> 
#> Iteration 25... 
#> Lower bound = -8611.57
#> 
#> Iteration 30... 
#> Lower bound = -8609.34
#> 
#> Iteration 35... 
#> Lower bound = -8608.655
#> 
#> Iteration 40... 
#> Lower bound = -8608.435
#> 
#> Iteration 45... 
#> Lower bound = -8608.362
#> 
#> Iteration 50... 
#> Lower bound = -8608.337
#> 
#> Iteration 55... 
#> Lower bound = -8608.328
#> 
#> Iteration 60... 
#> Lower bound = -8608.324
#> 
#> Iteration 65... 
#> Lower bound = -8608.323
#> 
#> Convergence obtained after 69 iterations with variational lower bound = -8608.323. 
#> 
#> ... done. == 
#> 

# Binary outcomes
#
dat_b <- generate_dependence(list_snps = list_X, list_phenos = list_Y,
                             ind_d0 = sample(1:d, d0),
                             ind_p0 = sample(1:p, p0), vec_prob_sh = 0.1,
                             family = "binomial", max_tot_pve = 0.9,
                             user_seed = user_seed)

list_hyper_logit <- set_hyper(d, p, lambda = 1, nu = 1, a = 1, b = 4*d-1,
                              eta = NULL, kappa = NULL,
                              link = "logit")

vb_logit <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0,
                  link = "logit", list_hyper = list_hyper_logit,
                  user_seed = user_seed)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -12848.19
#> 
#> Iteration 5... 
#> Lower bound = -5233.802
#> 
#> Iteration 10... 
#> Lower bound = -4480.087
#> 
#> Iteration 15... 
#> Lower bound = -4015.076
#> 
#> Iteration 20... 
#> Lower bound = -3606.86
#> 
#> Iteration 25... 
#> Lower bound = -3437.502
#> 
#> Iteration 30... 
#> Lower bound = -3313.61
#> 
#> Iteration 35... 
#> Lower bound = -3229.66
#> 
#> Iteration 40... 
#> Lower bound = -3211.858
#> 
#> Iteration 45... 
#> Lower bound = -3199.891
#> 
#> Iteration 50... 
#> Lower bound = -3186.358
#> 
#> Iteration 55... 
#> Lower bound = -3180.618
#> 
#> Iteration 60... 
#> Lower bound = -3170.698
#> 
#> Iteration 65... 
#> Lower bound = -3166.931
#> 
#> Iteration 70... 
#> Lower bound = -3164.682
#> 
#> Iteration 75... 
#> Lower bound = -3164.551
#> 
#> Iteration 80... 
#> Lower bound = -3164.514
#> 
#> Iteration 85... 
#> Lower bound = -3164.503
#> 
#> Iteration 90... 
#> Lower bound = -3164.5
#> 
#> Iteration 95... 
#> Lower bound = -3164.499
#> 
#> Convergence obtained after 98 iterations with variational lower bound = -3164.498. 
#> 
#> ... done. == 
#> 

list_hyper_probit <- set_hyper(d, p, lambda = 1, nu = 1, a = 1, b = 4*d-1,
                               eta = NULL, kappa = NULL,
                               link = "probit")

vb_probit <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0,
                   link = "probit", list_hyper = list_hyper_probit,
                   user_seed = user_seed)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -3376.773
#> 
#> Iteration 5... 
#> Lower bound = -3274.148
#> 
#> Iteration 10... 
#> Lower bound = -3243.948
#> 
#> Iteration 15... 
#> Lower bound = -3236.234
#> 
#> Iteration 20... 
#> Lower bound = -3230.538
#> 
#> Iteration 25... 
#> Lower bound = -3230.162
#> 
#> Iteration 30... 
#> Lower bound = -3230.133
#> 
#> Iteration 35... 
#> Lower bound = -3230.13
#> 
#> Convergence obtained after 38 iterations with variational lower bound = -3230.129. 
#> 
#> ... done. == 
#> 

# Binary outcomes with covariates
#
list_hyper_logit_z <- set_hyper(d, p, lambda = 1, nu = 1, a = 1, b = 4*d-1,
                                eta = NULL, kappa = NULL,
                                link = "logit", q = q, phi = 1,
                                xi = 1)

vb_logit_z <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0, Z = Z,
                    link = "logit",
                    list_hyper = list_hyper_logit_z, user_seed = user_seed)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -13359.45
#> 
#> Iteration 5... 
#> Lower bound = -5458.651
#> 
#> Iteration 10... 
#> Lower bound = -4742.715
#> 
#> Iteration 15... 
#> Lower bound = -4259.575
#> 
#> Iteration 20... 
#> Lower bound = -3897.018
#> 
#> Iteration 25... 
#> Lower bound = -3660.42
#> 
#> Iteration 30... 
#> Lower bound = -3520.332
#> 
#> Iteration 35... 
#> Lower bound = -3445.843
#> 
#> Iteration 40... 
#> Lower bound = -3358.993
#> 
#> Iteration 45... 
#> Lower bound = -3283.276
#> 
#> Iteration 50... 
#> Lower bound = -3259.899
#> 
#> Iteration 55... 
#> Lower bound = -3253.353
#> 
#> Iteration 60... 
#> Lower bound = -3247.442
#> 
#> Iteration 65... 
#> Lower bound = -3240.225
#> 
#> Iteration 70... 
#> Lower bound = -3237.613
#> 
#> Iteration 75... 
#> Lower bound = -3236.367
#> 
#> Iteration 80... 
#> Lower bound = -3236.231
#> 
#> Iteration 85... 
#> Lower bound = -3234.095
#> 
#> Iteration 90... 
#> Lower bound = -3234.05
#> 
#> Iteration 95... 
#> Lower bound = -3233.911
#> 
#> Iteration 100... 
#> Lower bound = -3233.363
#> 
#> Iteration 105... 
#> Lower bound = -3233.361
#> 
#> Convergence obtained after 106 iterations with variational lower bound = -3233.361. 
#> 
#> ... done. == 
#> 

list_hyper_probit_z <- set_hyper(d, p, lambda = 1, nu = 1, a = 1, b = 4*d-1,
                                 eta = NULL, kappa = NULL,
                                 link = "probit", q = q, phi = 1,
                                 xi = 1)

vb_probit_z <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0, Z = Z,
                     link = "probit",
                     list_hyper = list_hyper_probit_z, user_seed = user_seed)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -4801.484
#> 
#> Iteration 5... 
#> Lower bound = -3370.754
#> 
#> Iteration 10... 
#> Lower bound = -3356.969
#> 
#> Iteration 15... 
#> Lower bound = -3352.259
#> 
#> Iteration 20... 
#> Lower bound = -3351.955
#> 
#> Iteration 25... 
#> Lower bound = -3351.934
#> 
#> Iteration 30... 
#> Lower bound = -3351.93
#> 
#> Convergence obtained after 35 iterations with variational lower bound = -3351.929. 
#> 
#> ... done. == 
#> 

# Mix of continuous and binary outcomes
#
Y_mix <- cbind(dat_g$phenos, dat_b$phenos)
ind_bin <- (d+1):(2*d)
p0_mix <- sum(rowSums(cbind(dat_g$pat, dat_b$pat)) > 0)

list_hyper_mix <- set_hyper(2*d, p, lambda = 1, nu = 1, a = 1, b = 8*d-1,
                            eta = 1, kappa = apply(dat_g$phenos, 2, var),
                            link = "mix", ind_bin = ind_bin)

vb_mix <- locus(Y = Y_mix, X = dat_b$snps, p0_av = p0_mix, link = "mix",
                ind_bin = ind_bin, list_hyper = list_hyper_mix,
                user_seed = user_seed)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -13491.11
#> 
#> Iteration 5... 
#> Lower bound = -11170.93
#> 
#> Iteration 10... 
#> Lower bound = -11137.25
#> 
#> Iteration 15... 
#> Lower bound = -11127.95
#> 
#> Iteration 20... 
#> Lower bound = -11126.82
#> 
#> Iteration 25... 
#> Lower bound = -11123.58
#> 
#> Iteration 30... 
#> Lower bound = -11117.02
#> 
#> Iteration 35... 
#> Lower bound = -11114.66
#> 
#> Iteration 40... 
#> Lower bound = -11111.29
#> 
#> Iteration 45... 
#> Lower bound = -11110.68
#> 
#> Iteration 50... 
#> Lower bound = -11110.64
#> 
#> Iteration 55... 
#> Lower bound = -11110.64
#> 
#> Iteration 60... 
#> Lower bound = -11110.64
#> 
#> Convergence obtained after 65 iterations with variational lower bound = -11110.64. 
#> 
#> ... done. == 
#> 

list_hyper_mix_z <- set_hyper(2*d, p, lambda = 1, nu = 1, a = 1, b = 8*d-1,
                              eta = 1, kappa = apply(dat_g$phenos, 2, var),
                              link = "mix", ind_bin = ind_bin, q = q,
                              phi = 1, xi = 1)

vb_mix_z <- locus(Y = Y_mix, X = dat_b$snps, p0_av = p0_mix, Z = Z,
                  link = "mix", ind_bin = ind_bin,
                  list_hyper = list_hyper_mix_z, user_seed = user_seed)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> Seed set to user_seed 123. 
#> list_init set automatically. 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -16608.99
#> 
#> Iteration 5... 
#> Lower bound = -11372.94
#> 
#> Iteration 10... 
#> Lower bound = -11352.12
#> 
#> Iteration 15... 
#> Lower bound = -11339.49
#> 
#> Iteration 20... 
#> Lower bound = -11336.35
#> 
#> Iteration 25... 
#> Lower bound = -11327.61
#> 
#> Iteration 30... 
#> Lower bound = -11326.77
#> 
#> Iteration 35... 
#> Lower bound = -11326.68
#> 
#> Iteration 40... 
#> Lower bound = -11326.67
#> 
#> Iteration 45... 
#> Lower bound = -11326.67
#> 
#> Convergence obtained after 49 iterations with variational lower bound = -11326.67. 
#> 
#> ... done. == 
#>