Gather initial variational parameters provided by the user.

This function must be used to provide initial values for the variational parameters used in locus.

set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb, tau_vb, link = "identity",
  ind_bin = NULL, q = NULL, mu_alpha_vb = NULL, sig2_alpha_vb = NULL,
  r = NULL, mu_c0_vb = NULL, mu_c_vb = NULL)

Arguments

d
Number of responses.

p
Number of candidate predictors.

gam_vb
Matrix of size p x d with initial values for the variational parameter yielding posterior probabilities of inclusion.

mu_beta_vb
Matrix of size p x d with initial values for the variational parameter yielding regression coefficient estimates for predictor-response pairs included in the model.

sig2_beta_vb
Vector of length d, for link = "identity" and for link = "mix", of length 1 for link = "probit", and a matrix of size p x d, for link = "logit", with initial values for the variational parameter yielding estimates of effect variances for predictor-response pairs included in the model. For link = "identity" and link = "mix", these values are the same for all the predictors (as a result of the predictor variables being standardized before the variational algorithm). For link = "probit", they are the same for all the predictors and responses.

tau_vb
Vector of length d, for link = "identity", and of length d_cont = d - length(ind_bin) (number of continuous responses), for link = "mix", with initial values for the variational parameter yielding estimates for the continuous response residual precisions. Must be NULL for link = "logit" and link = "probit".

link
Response link. Must be "identity" for linear regression, "logit" for logistic regression, "probit" for probit regression, or "mix" for a mix of identity and probit link functions (in this case, the indices of the binary responses must be gathered in argument ind_bin, see below).

ind_bin
If link = "mix", vector of indices corresponding to the binary variables in Y. Must be NULL if link != "mix".

q
Number of covariates. Default is NULL, for Z NULL.

mu_alpha_vb
Matrix of size q x d with initial values for the variational parameter yielding regression coefficient estimates for covariate-response pairs. Default is NULL, for Z NULL.

sig2_alpha_vb
Matrix of size q x d for link = "identity", for link = "logit" and for link = "mix" with initial values for the variational parameter yielding estimates of effect variances for covariate-response pairs. Vector of length q for link = "probit". Default is NULL, for Z NULL.

r
Number of variables representing external information on the candidate predictors. Default is NULL, for V NULL.

mu_c0_vb
Vector of length p with initial values for the variational parameter linked to the proportion of responses associated with each candidate predictor. Default is NULL, for V NULL.

mu_c_vb
Matrix of size r x d with initial values for the variational parameter yielding regression coefficient estimates for the influence of external information on the candidate predictors on their selection. Default is NULL, for V NULL.

Value

An object of class "init" preparing user initial values for the variational parameters in a form that can be passed to the locus function.

Details

The locus function can also be used with default initial parameter choices (without using set_init) by setting its argument list_init to NULL.

See also

set_hyper, locus

Examples

user_seed <- 123; set.seed(user_seed)
n <- 200; p <- 250; p0 <- 50; d <- 25; d0 <- 15
list_X <- generate_snps(n = n, p = p)
list_Y <- generate_phenos(n = n, d = d, var_err = 1)

# Continuous outcomes
#
dat_g <- generate_dependence(list_snps = list_X, list_phenos = list_Y,
                             ind_d0 = sample(1:d, d0),
                             ind_p0 = sample(1:p, p0),
                             vec_prob_sh = 0.1, family = "gaussian",
                             max_tot_pve = 0.9)

# gam_vb chosen so that the prior mean number of responses associated with
# each candidate predictor is 1/4.
gam_vb <- matrix(rbeta(p * d, shape1 = 1, shape2 = 4*d-1), nrow = p)
mu_beta_vb <- matrix(rnorm(p * d), nrow = p)
tau_vb <- 1 / apply(dat_g$phenos, 2, var)
sig2_beta_vb <- 1 / rgamma(d, shape = 2, rate = 1 / tau_vb)

list_init_g <- set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb, tau_vb,
                        link = "identity")

vb_g <- locus(Y = dat_g$phenos, X = dat_g$snps, p0_av = p0,
              link = "identity", list_init = list_init_g)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -8392.149
#> 
#> Iteration 5... 
#> Lower bound = -7732.243
#> 
#> Iteration 10... 
#> Lower bound = -7731.908
#> 
#> Convergence obtained after 12 iterations with variational lower bound = -7731.908. 
#> 
#> ... done. == 
#> 

# Continuous outcomes with covariates
#
q <- 4
Z <- matrix(rnorm(n * q), nrow = n)

mu_alpha_vb <- matrix(rnorm(q * d), nrow = q)
sig2_alpha_vb <- 1 / matrix(rgamma(q * d, shape = 2, rate = 1), nrow = q)

list_init_g_z <- set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb, tau_vb,
                          link = "identity", q = q,
                          mu_alpha_vb = mu_alpha_vb,
                          sig2_alpha_vb = sig2_alpha_vb)

# we take p0_av = p0 (known here); this choice may result in variable
# selections that are (too) conservative in some cases. In practice, often
# p0_av as a slightly overestimated guess of p0.
vb_g_z <- locus(Y = dat_g$phenos, X = dat_g$snps, p0_av = p0, Z = Z,
                link = "identity", list_init = list_init_g_z)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -9010.233
#> 
#> Iteration 5... 
#> Lower bound = -7864.928
#> 
#> Iteration 10... 
#> Lower bound = -7863.425
#> 
#> Convergence obtained after 13 iterations with variational lower bound = -7863.424. 
#> 
#> ... done. == 
#> 

# Continuous outcomes with external annotation
#
r <- 4
V <- matrix(rnorm(p * r), nrow = p)
bool_p0 <- rowSums(dat_g$pat) > 0
V[bool_p0, ] <- rnorm(sum(bool_p0) * r, mean = 2) # informative annotations

mu_c0_vb <- rnorm(p, mean = -1)
mu_c_vb <- matrix(rnorm(r * d, mean = 0, sd = 0.01), nrow = r)

list_init_g_v <- set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb, tau_vb,
                          link = "identity", r = r, mu_c0_vb = mu_c0_vb,
                          mu_c_vb = mu_c_vb)

vb_g_v <- locus(Y = dat_g$phenos, X = dat_g$snps, p0_av = p0,  V = V,
                link = "identity", list_init = list_init_g_v)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -22253.94
#> 
#> Iteration 5... 
#> Lower bound = -18051.88
#> 
#> Iteration 10... 
#> Lower bound = -15282.19
#> 
#> Iteration 15... 
#> Lower bound = -13413.59
#> 
#> Iteration 20... 
#> Lower bound = -12136.17
#> 
#> Iteration 25... 
#> Lower bound = -11270.42
#> 
#> Iteration 30... 
#> Lower bound = -10662.13
#> 
#> Iteration 35... 
#> Lower bound = -10248.23
#> 
#> Iteration 40... 
#> Lower bound = -9964.417
#> 
#> Iteration 45... 
#> Lower bound = -9768.09
#> 
#> Iteration 50... 
#> Lower bound = -9634.518
#> 
#> Iteration 55... 
#> Lower bound = -9544.132
#> 
#> Iteration 60... 
#> Lower bound = -9481.631
#> 
#> Iteration 65... 
#> Lower bound = -9439.442
#> 
#> Iteration 70... 
#> Lower bound = -9411.242
#> 
#> Iteration 75... 
#> Lower bound = -9392.295
#> 
#> Iteration 80... 
#> Lower bound = -9379.538
#> 
#> Iteration 85... 
#> Lower bound = -9370.94
#> 
#> Iteration 90... 
#> Lower bound = -9365.14
#> 
#> Iteration 95... 
#> Lower bound = -9361.224
#> 
#> Iteration 100... 
#> Lower bound = -9358.571
#> 
#> Iteration 105... 
#> Lower bound = -9356.651
#> 
#> Iteration 110... 
#> Lower bound = -9336.916
#> 
#> Iteration 115... 
#> Lower bound = -9320.643
#> 
#> Iteration 120... 
#> Lower bound = -9312.615
#> 
#> Iteration 125... 
#> Lower bound = -9307.823
#> 
#> Iteration 130... 
#> Lower bound = -9304.777
#> 
#> Iteration 135... 
#> Lower bound = -9302.798
#> 
#> Iteration 140... 
#> Lower bound = -9301.497
#> 
#> Iteration 145... 
#> Lower bound = -9300.635
#> 
#> Iteration 150... 
#> Lower bound = -9300.06
#> 
#> Iteration 155... 
#> Lower bound = -9299.675
#> 
#> Iteration 160... 
#> Lower bound = -9299.416
#> 
#> Iteration 165... 
#> Lower bound = -9299.24
#> 
#> Iteration 170... 
#> Lower bound = -9299.113
#> 
#> Iteration 175... 
#> Lower bound = -9298.742
#> 
#> Iteration 180... 
#> Lower bound = -9293.966
#> 
#> Iteration 185... 
#> Lower bound = -9290.697
#> 
#> Iteration 190... 
#> Lower bound = -9288.476
#> 
#> Iteration 195... 
#> Lower bound = -9286.367
#> 
#> Iteration 200... 
#> Lower bound = -9275.842
#> 
#> Iteration 205... 
#> Lower bound = -9259.815
#> 
#> Iteration 210... 
#> Lower bound = -9251.461
#> 
#> Iteration 215... 
#> Lower bound = -9245.98
#> 
#> Iteration 220... 
#> Lower bound = -9242.297
#> 
#> Iteration 225... 
#> Lower bound = -9239.814
#> 
#> Iteration 230... 
#> Lower bound = -9238.138
#> 
#> Iteration 235... 
#> Lower bound = -9237.007
#> 
#> Iteration 240... 
#> Lower bound = -9236.243
#> 
#> Iteration 245... 
#> Lower bound = -9235.727
#> 
#> Iteration 250... 
#> Lower bound = -9235.379
#> 
#> Iteration 255... 
#> Lower bound = -9235.143
#> 
#> Iteration 260... 
#> Lower bound = -9234.984
#> 
#> Iteration 265... 
#> Lower bound = -9234.877
#> 
#> Iteration 270... 
#> Lower bound = -9234.804
#> 
#> Iteration 275... 
#> Lower bound = -9234.755
#> 
#> Iteration 280... 
#> Lower bound = -9234.722
#> 
#> Iteration 285... 
#> Lower bound = -9234.7
#> 
#> Iteration 290... 
#> Lower bound = -9234.685
#> 
#> Iteration 295... 
#> Lower bound = -9234.675
#> 
#> Iteration 300... 
#> Lower bound = -9234.668
#> 
#> Iteration 305... 
#> Lower bound = -9234.663
#> 
#> Iteration 310... 
#> Lower bound = -9234.66
#> 
#> Iteration 315... 
#> Lower bound = -9234.658
#> 
#> Iteration 320... 
#> Lower bound = -9234.656
#> 
#> Iteration 325... 
#> Lower bound = -9234.655
#> 
#> Iteration 330... 
#> Lower bound = -9234.655
#> 
#> Convergence obtained after 333 iterations with variational lower bound = -9234.654. 
#> 
#> ... done. == 
#> 

# Binary outcomes
#
dat_b <- generate_dependence(list_snps = list_X, list_phenos = list_Y,
                             ind_d0 = sample(1:d, d0),
                             ind_p0 = sample(1:p, p0),
                             vec_prob_sh = 0.1, family = "binomial",
                             max_tot_pve = 0.9)

# gam_vb chosen so that the prior mean number of responses associated with
# each candidate predictor is 1/4.
sig2_beta_vb_logit <- 1 / t(replicate(p, rgamma(d, shape = 2, rate = 1)))

list_init_logit <- set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb_logit,
                            tau_vb = NULL, link = "logit")

vb_logit <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0,
                  link = "logit", list_init = list_init_logit)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -3408.328
#> 
#> Iteration 5... 
#> Lower bound = -3091.242
#> 
#> Iteration 10... 
#> Lower bound = -3088.856
#> 
#> Iteration 15... 
#> Lower bound = -3088.021
#> 
#> Iteration 20... 
#> Lower bound = -3087.633
#> 
#> Iteration 25... 
#> Lower bound = -3087.617
#> 
#> Convergence obtained after 30 iterations with variational lower bound = -3087.616. 
#> 
#> ... done. == 
#> 

sig2_beta_vb_probit <- sig2_beta_vb[1]
list_init_probit <- set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb_probit,
                             tau_vb = NULL, link = "probit")

vb_probit <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0,
                   link = "probit", list_init = list_init_probit)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -3270.045
#> 
#> Iteration 5... 
#> Lower bound = -3141.964
#> 
#> Iteration 10... 
#> Lower bound = -3129.708
#> 
#> Iteration 15... 
#> Lower bound = -3128.187
#> 
#> Iteration 20... 
#> Lower bound = -3128.163
#> 
#> Iteration 25... 
#> Lower bound = -3128.162
#> 
#> Convergence obtained after 26 iterations with variational lower bound = -3128.162. 
#> 
#> ... done. == 
#> 

# Binary outcomes with covariates
#
list_init_logit_z <- set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb_logit,
                              tau_vb = NULL, link = "logit",
                              q = q, mu_alpha_vb = mu_alpha_vb,
                              sig2_alpha_vb = sig2_alpha_vb)

vb_logit_z <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0, Z = Z,
                   link = "logit", list_init = list_init_logit_z)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -4201.992
#> 
#> Iteration 5... 
#> Lower bound = -3169.318
#> 
#> Iteration 10... 
#> Lower bound = -3167.883
#> 
#> Iteration 15... 
#> Lower bound = -3167.778
#> 
#> Iteration 20... 
#> Lower bound = -3167.729
#> 
#> Convergence obtained after 24 iterations with variational lower bound = -3167.729. 
#> 
#> ... done. == 
#> 

sig2_alpha_vb_probit <- sig2_alpha_vb[, 1]
list_init_probit_z <- set_init(d, p, gam_vb, mu_beta_vb, sig2_beta_vb_probit,
                               tau_vb = NULL, link = "probit",
                               q = q, mu_alpha_vb = mu_alpha_vb,
                               sig2_alpha_vb = sig2_alpha_vb_probit)

vb_probit_z <- locus(Y = dat_b$phenos, X = dat_b$snps, p0_av = p0, Z = Z,
                     link = "probit", list_init = list_init_probit_z)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 50 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -4474.302
#> 
#> Iteration 5... 
#> Lower bound = -3283.438
#> 
#> Iteration 10... 
#> Lower bound = -3271.126
#> 
#> Iteration 15... 
#> Lower bound = -3270.917
#> 
#> Iteration 20... 
#> Lower bound = -3268.928
#> 
#> Iteration 25... 
#> Lower bound = -3268.884
#> 
#> Convergence obtained after 28 iterations with variational lower bound = -3268.884. 
#> 
#> ... done. == 
#> 

# Mix of continuous and binary outcomes
#
Y_mix <- cbind(dat_g$phenos, dat_b$phenos)
ind_bin <- (d+1):(2*d)
p0_mix <- sum(rowSums(cbind(dat_g$pat, dat_b$pat)) > 0)

# gam_vb chosen so that the prior mean number of responses associated with
# each candidate predictor is 1/4.
gam_vb_mix <- matrix(rbeta(p * 2*d, shape1 = 1, shape2 = 8*d-1), nrow = p)
mu_beta_vb_mix <- matrix(rnorm(p * 2*d), nrow = p)
sig2_beta_vb_mix <- 1 / c(rgamma(d, shape = 2, rate = 1 / tau_vb),
                          rgamma(d, shape = 2, rate = 1))

list_init_mix <- set_init(2*d, p, gam_vb_mix, mu_beta_vb_mix,
                          sig2_beta_vb_mix, tau_vb, link = "mix",
                          ind_bin = ind_bin)

vb_mix <- locus(Y = Y_mix, X = dat_b$snps, p0_av = p0_mix, link = "mix",
                ind_bin = ind_bin, list_init = list_init_mix)

#> == Preparing the data ... 
#> ... done. == 
#> 
#> Provided p0_av = 89 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -11689.61
#> 
#> Iteration 5... 
#> Lower bound = -10911.23
#> 
#> Iteration 10... 
#> Lower bound = -10900.25
#> 
#> Iteration 15... 
#> Lower bound = -10898.35
#> 
#> Iteration 20... 
#> Lower bound = -10898.33
#> 
#> Convergence obtained after 24 iterations with variational lower bound = -10898.33. 
#> 
#> ... done. == 
#> 

mu_alpha_vb_mix <- matrix(rnorm(q * 2*d), nrow = q)
sig2_alpha_vb_mix <- 1 / matrix(rgamma(q * 2*d, shape = 2, rate = 1), nrow = q)

list_init_mix_z <- set_init(2*d, p, gam_vb_mix, mu_beta_vb_mix,
                            sig2_beta_vb_mix, tau_vb, link = "mix",
                            ind_bin = ind_bin, q = q,
                            mu_alpha_vb = mu_alpha_vb_mix,
                            sig2_alpha_vb = sig2_alpha_vb_mix)

vb_mix_z <- locus(Y = Y_mix, X = dat_b$snps, p0_av = p0_mix, Z = Z,
                  link = "mix", ind_bin = ind_bin,
                  list_init = list_init_mix_z)

#> == Preparing the data ... 
#> Each factor variables must be provided by adding to Z (nb levels - 1) variables representing their levels. 
#> ... done. == 
#> 
#> Provided p0_av = 89 interpreted as the prior number of predictors associated with at least one response. 
#> 
#> == Preparing the hyperparameters ... 
#> 
#> list_hyper set automatically. 
#> ... done. == 
#> 
#> == Preparing the parameter initialization ... 
#> 
#> ... done. == 
#> 
#> ============================================================== 
#> == Variational inference for sparse multivariate regression == 
#> ============================================================== 
#> 
#> Iteration 1... 
#> Lower bound = -13583.81
#> 
#> Iteration 5... 
#> Lower bound = -11152.01
#> 
#> Iteration 10... 
#> Lower bound = -11116.51
#> 
#> Iteration 15... 
#> Lower bound = -11112.56
#> 
#> Iteration 20... 
#> Lower bound = -11110.39
#> 
#> Iteration 25... 
#> Lower bound = -11110.38
#> 
#> Iteration 30... 
#> Lower bound = -11110.38
#> 
#> Convergence obtained after 32 iterations with variational lower bound = -11110.38. 
#> 
#> ... done. == 
#>